refactor: split login screen into separate file
